added more entities

+++ src/main/java/com/example/is442oop/vmsbackend/entities/Form.java
@@ -5,13 +5,13 @@ import java.util.*;
 @Entity
 public class Form {
     @Id
-    private Integer formID;
+    private Long formID;
     private String HTML;
     @OneToMany(mappedBy="form", cascade=CascadeType.ALL, fetch=FetchType.LAZY)
     public Set <Questionnaire> questionnaire;
 
     
-    public Form(Integer formID, String hTML) {
+    public Form(Long formID, String hTML) {
         this.formID = formID;
         HTML = hTML;
     }
@@ -21,10 +21,10 @@ public class Form {
     }
 
 
-    public Integer getFormID() {
+    public Long getFormID() {
         return formID;
     }
-    public void setFormID(Integer formID) {
+    public void setFormID(Long formID) {
         this.formID = formID;
     }
     public String getHTML() {
